Remove optional Azure connection from Create-ResourceGroup.ps1
The parameter that controlled the option to connect to Azure has been removed from the Create-ResourceGroup.ps1 script. Consequently, the script will always import the Connect-Azure module and establish a connection with Azure using the provided SubscriptionId. This change simplifies the flow by always ensuring a connection is established when creating a new resource group.