Refactor Connect-Azure module and add New-RandomString function
The Connect-Azure module has been renamed to Tools and refactored to improve the error handling. The refactoring includes a change in how parameters for Connect-AzAccount are passed and how the federated token is read. In addition, the New-RandomString function has been added, expanding the module's functionality to provide a random string based on specified conditions.