Rearrange parameter order in Azure module and workflow
The parameter order in the powershell Azure module and the new-storage-account.yaml workflow file has been rearranged. The 'CreateResourceGroup' parameter is moved after the 'ResourceGroupName' parameter. This change aims to provide a more logical flow by asking for the resource group name first and then asking whether to create that resource group.