Add location-based abbreviation to VM prefix
The code has been updated to include a location-based abbreviation in the virtual machine prefix for more accurate identification. Previously, all VMs in the Azure module had a prefix "azneulx", now it is dynamically set based on the location (e.g., "azneulx" for North Europe, "azweulx" for West Europe, and "aznalx" for other locations).